Checking how git detect changes
Putting a space at top to see how git detect changes. Does it just match the code on each line (i.e line having same number) and a blank line can have a huge impact on the code or is it smart enough to overlook it and is able to make out that there is no change in code.